v1.40.40 Released!
New exchange Poloniex
Fixed script language command cancel ask and bid for Binance
Fixed bug caused forever API down in some cases
Minor fixes
Drop support of macOS 10.11

+++ src/qtbitcointrader.h
@@ -1,6 +1,6 @@
 //  This file is part of Qt Bitcoin Trader
 //      https://github.com/JulyIGHOR/QtBitcoinTrader
-//  Copyright (C) 2013-2018 July IGHOR <julyighor@gmail.com>
+//  Copyright (C) 2013-2019 July Ighor <julyighor@gmail.com>
 //
 //  This program is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU General Public License as published by
@@ -135,7 +135,7 @@ public:
 
     OrdersModel* ordersModel;
 
-    quint32 currencyChangedDate;
+    qint64 currencyChangedDate;
 
     QSettings* iniSettings;
     bool isValidSoftLag;
